Moved DbContext core logic from InvestorData project to InvestorApi.Data project, leaving Models for client usage.
